reorganize receive_and_buffer to be shared with Scheduler bench
